sync-github-releases: extract git & CI-event queries into git.ts
The functions that shell out to git (repo root, tag existence, release-commit
lookup, tracked/changed CHANGELOG.md files) and read the GitHub Actions event to
decide which package dirs to sync are one concern: querying the local repo and
CI context. Move them out of index.ts, which is now just the entry point and the
syncPackage orchestration.
