+++ .github/workflows/sync-github-releases/index.spec.ts
@@ -2,8 +2,8 @@ import { readFileSync } from 'node:fs'
 import path from 'node:path'
 import { describe, expect, it } from 'vitest'
 import { parseChangelog } from './changelog.ts'
+import { toPackageDirs } from './git.ts'
 import { getRepository } from './github-utils.ts'
-import { toPackageDirs } from './index.ts'
 import { chooseCreateCommitish, getReleasePlan, getTagName, withSourceOfTruth } from './release-plan.ts'
 
 function readFixture(name: string): string {
